Big update to 0.3.  Lots of fixes
This is a big update as it makes changes to how things are saved.  It's a lot more stable (though still a bit buggy, but not crashing/laggy buggy).  Lots of fixes to logging and other misc stuff.  Theres a lot of changes so old watchlists may not work now.  Export, import should work now.  Popup fixes.